refactor(backend): drop extension forwarding shell (#876)
