+++ docs/a2a-consumer-baseline.md
@@ -93,16 +93,13 @@ Not allowed:
 
 ## Current Repository Abstraction Inventory
 
-The repository still contains a small number of consumer-side abstractions. They
-should not all be treated the same during follow-up cleanup.
+The repository still contains a small number of consumer-side abstractions. They should not all be treated the same during follow-up cleanup.
 
 ### 1. Upstream Protocol Facts
 
-These values remain upstream-owned facts even when they contain provider or
-family-specific wording:
+These values remain upstream-owned facts even when they contain provider or family-specific wording:
 
-- upstream JSON-RPC method literals such as `codex.discovery.*`,
-  `codex.threads.*`, `codex.turns.*`, `codex.review.*`, and `codex.exec.*`
+- upstream JSON-RPC method literals such as `codex.discovery.*`, `codex.threads.*`, `codex.turns.*`, `codex.review.*`, and `codex.exec.*`
 - declared metadata fields such as `metadata.codex.execution`
 - supported extension URI families that the Hub intentionally recognizes
 
@@ -114,8 +111,7 @@ Decision rule:
 
 ### 2. Stable Public Hub Contracts
 
-Some names are no longer merely implementation details because they are already
-part of the backend/frontend contract exposed by the Hub:
+Some names are no longer merely implementation details because they are already part of the backend/frontend contract exposed by the Hub:
 
 - `upstreamMethodFamilies`
 - `providerDiscovery`
@@ -129,22 +125,17 @@ Decision rule:
 
 ### 3. Repo-Local Adapter Abstractions
 
-These are the main candidates for continued cleanup because they exist to keep
-the adapter layer operable, not because the upstream protocol requires them:
+These are the main candidates for continued cleanup because they exist to keep the adapter layer operable, not because the upstream protocol requires them:
 
-- adapter-scoped metadata routing hints such as backend-only metadata
-  namespaces
-- adapter-local metadata field inventories used by session binding and model
-  selection resolvers
-- narrow compatibility helpers that translate declared contracts into Hub
-  runtime inputs without changing protocol meaning
+- adapter-scoped metadata routing hints such as backend-only metadata namespaces
+- adapter-local metadata field inventories used by session binding and model selection resolvers
+- narrow compatibility helpers that translate declared contracts into Hub runtime inputs without changing protocol meaning
 
 Decision rule:
 
 - prefer neutral adapter wording over provider-private wording
 - keep them backend-local whenever possible
-- remove them when the upstream contract no longer needs explicit adapter
-  routing
+- remove them when the upstream contract no longer needs explicit adapter routing
 
 ## ACP Reference Boundary
 
@@ -153,8 +144,7 @@ ACP is intentionally not part of the current implementation baseline.
 Current rule:
 
 - ACP may be tracked as a later comparison input for protocol-boundary review
-- ACP should not drive naming or contract changes in the current A2A consumer
-  cleanup until a dedicated evaluation is opened
+- ACP should not drive naming or contract changes in the current A2A consumer cleanup until a dedicated evaluation is opened
 
 ## `card:validate` Responsibility
